+++ hyperdbg/hprdbgctrl/header/common.h
@@ -81,6 +81,21 @@
 #define CPUID_ADDR_WIDTH 0x80000008
 
 //////////////////////////////////////////////////
+//			  Assembly Functions                //
+//////////////////////////////////////////////////
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+extern bool
+AsmVmxSupportDetection();
+
+#ifdef __cplusplus
+}
+#endif
+
+//////////////////////////////////////////////////
 //			    	 Spinlocks                  //
 //////////////////////////////////////////////////
 
@@ -98,7 +113,7 @@ SpinlockUnlock(volatile LONG * Lock);
 //////////////////////////////////////////////////
 
 VOID
-PrintBits(size_t const size, void const * const ptr);
+PrintBits(const UINT32 size, const void * ptr);
 
 BOOL
 Replace(std::string & str, const std::string & from, const std::string & to);
@@ -133,13 +148,10 @@ HasEnding(string const & fullString, string const & ending);
 BOOLEAN
 ValidateIP(const string & ip);
 
-BOOLEAN
-VmxSupportDetection();
-
 BOOL
-SetPrivilege(HANDLE  hToken,          // access token handle
-             LPCTSTR lpszPrivilege,   // name of privilege to enable/disable
-             BOOL    bEnablePrivilege // to enable or disable privilege
+SetPrivilege(HANDLE  Token,          // access token handle
+             LPCTSTR Privilege,      // name of privilege to enable/disable
+             BOOL    EnablePrivilege // to enable or disable privilege
 );
 
 void