+++ hyperdbg/hprdbgctrl/header/list.h
@@ -11,5 +11,7 @@
  */
 #pragma once
 
+/*
 #define CONTAINING_RECORD(address, type, field) \
     ((type *)((char *)(address) - (unsigned long)(&((type *)0)->field)))
+*/