+++ CHANGELOG.md
@@ -8,14 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 New release of the HyperDbg Debugger.
 
 ### Added
-- Add user-defined functions and variable types in script engine thanks to [@xmaple555](https://github.com/xmaple555) ([link](https://github.com/HyperDbg/HyperDbg/pull/342))
+- Add user-defined functions and variable types in script engine thanks to [@xmaple555](https://github.com/xmaple555) ([link](https://docs.hyperdbg.org/commands/scripting-language/constants-and-functions))([link](https://github.com/HyperDbg/HyperDbg/pull/342))
 
 ### Changed
 - Fix debuggee crash after running the '.debug close' command on the debugger  
 - After downloading new symbols it is automatically loaded
 - The problem with adding edge MTRR pages is fixed thanks to [@Maladiy](https://github.com/Maladiy) ([link](https://github.com/HyperDbg/HyperDbg/pull/347)) 
 - All compiler/linker warnings of kernel-mode modules are fixed  
-- Kernel modules of HyperDbg now compile with "treat warning as error"  
+- User/Kernel modules of HyperDbg now compile with "treat warning as error"  
 
 ## [0.8.1.0] - 2024-02-01
 New release of the HyperDbg Debugger.